Allow output verbosity below normal
Since the cli verbosity is 'normal' most of the time using max was not
working if you want to use 'quiet' in your configuration file.

Now cli verbosity is ignored as long as it is normal. If it is set
manually, it overwrites the configuration file value.

So it is cli > config (but ignore normal)

Fixes #276